# [GR-45654] Single compile unit for Java DWARF debug info