--- a/substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/debugentry/DirEntry.java
+++ b/substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/debugentry/DirEntry.java
@@ -37,9 +37,11 @@ import java.nio.file.Path;
  */
 public class DirEntry {
     private Path path;
+    private int idx;
 
-    public DirEntry(Path path) {
+    public DirEntry(Path path, int idx) {
         this.path = path;
+        this.idx = idx;
     }
 
     public Path getPath() {
@@ -49,4 +51,14 @@ public class DirEntry {
     public String getPathString() {
         return path.toString();
     }
+
+    /**
+     * Retrieve the index of the dir entry in the list of all known dirs
+     *
+     * @return the index of the file entry.
+     */
+    public int getIdx() {
+        return idx;
+    }
+
 }