This PR reorganizes the DWARF debug info layout so that all info is all contained in a single compile unit (CU) rather than employing a CU for every Java class type.

It profits from that reorganization to achieve several important goals:
 - reduce the overall size of debug info by avoiding replication of DIEs shared by multiple classes
 - reduce the overall size of debug info by omitting unnecessary CU boiler plate
 - decouple compiled method info and definition order from class declaration order, thereby freeing the image builder to order method code ranges without respect to class grouping

The re-implementation also has the benefit that it simplifies many of the DWARF generation steps, both when iterating the generic debug model to write specific debug sections and when cross-indexing location references within and across sections.